requirements.txt
fastapi
uvicorn[standard]
python-multipart

Pillow

torch
torchvision